// src/app/models/billing/ORMUserInsert.ts
export class ORMUserInsert {
    user_id: String;
    user_name: String;
    password: String;
    first_name: String;
    mname: String;
    last_name: String;
    user_type: String;
    is_provider: Boolean;
    is_blocked: Boolean;
    signature_path: String;
    practice_id: String;
    provider_id: String;
    default_physician: String;
    default_billing_phy: String;
    default_role: String;
    default_location: String;
    default_bill: String;
    deleted: Boolean;
    created_user: String;
    modified_user: String;
    client_date_created: String;
    client_date_modified: String;
    date_created: String;
    date_modified: String;
    default_chart_setting: String;
    default_prescription_role: String;
    lab_assigned: Boolean;
    is_lab_user: Boolean;
}